+++ config.py
@@ -1,0 +1,9 @@
+"""配置文件"""
+URL = "http://39.106.82.121/query/getStudentScore"  # 请求地址
+DELAY = 0.1  # 每次请求间隔时间，单位秒
+TIMEOUT = 10
+
+WEEK_NUM = 20  # 学期的周数，默认为20
+
+BUILDINGS = ["1N", "1S", "2N", "2S", "3S", "4N", "4S", "5N",
+             "5S", "6N", "6S", "7#", "8#", "9N", "9S", "10N", "10S"]
